requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
